requirements.txt
opencv-python==4.8.1.78
dlib==19.24.99
face-recognition==1.3.0
mediapipe==0.9.0.1
norfair==2.2.0
tensorflow==2.4.0
torch==1.6.0
gdown==4.7.1
tqdm==4.66.1
python-dotenv
wandb==0.13.8